combined modals + fixes
