+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cs3api4lab",
+  "name": "@sciencemesh/cs3api4lab",
   "version": "0.1.0",
   "description": "CS3APIs Jupyterlab contents connector",
   "keywords": [
@@ -59,6 +59,7 @@
     "@jupyterlab/docmanager": "^3.0.11",
     "@jupyterlab/statedb": "^3.0.6",
     "@lumino/widgets": "^1.16.1",
+    "jupyterlab_toastify": "^4.2.1",
     "moment": "^2.27.0",
     "react-dropdown-select": "^4.7.4"
   },
